requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
python-jose[cryptography]==3.5.0
cryptography==46.0.5
pytest==9.0.2
pytest-cov==7.0.0
httpx==0.28.1